+++ src/io/import/main.test.ts
@@ -1,204 +1,75 @@
-import { importXmlFiles } from './main'
-import { handleExpectedRecords } from './test.handler'
-
-import Dexie from 'dexie'
-import { describe, expect, it, afterAll } from 'vitest'
+import { describe } from 'vitest'
 
 import { CUSTOM_RECORD_ID_ATTRIBUTE } from '@/helpers'
+import { importXmlFiles } from '@/io/import'
 import {
-	DIALECTE_NAMESPACES,
 	XMLNS_EXT_NAMESPACE,
 	TEST_DIALECTE_CONFIG,
 	XMLNS_DEFAULT_NAMESPACE,
 	XMLNS_DEV_NAMESPACE,
+	runTestCases,
 } from '@/test'
 
-import type { AnyDatabaseInstance } from '../database'
-import type { ExpectedRecords } from './test.types'
+import type { BaseXmlTestCase, ActResult } from '@/test'
 
 describe('Import', () => {
 	describe('Feature', () => {
-		const databaseNames: string[] = []
+		type TestCase = BaseXmlTestCase
 
-		afterAll(async () => {
-			for (const dbName of databaseNames) {
-				await Dexie.delete(dbName)
-			}
-		})
-
-		type TestCase = {
-			description: string
-			fileContent: string
-			expectedFileName: string
-			expectedRecords: ExpectedRecords
-		}
-
-		const testCases: TestCase[] = [
-			{
-				description: 'empty root',
-				fileContent: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"></Root>`,
-				expectedFileName: 'empty-root',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-					},
-				],
+		const testCases: Record<string, TestCase> = {
+			'empty root': {
+				sourceXml: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"></Root>`,
+				expectedQueries: ['//default:Root'],
 			},
-			{
-				description: 'child with attribute',
-				fileContent: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"><A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A></Root>`,
-				expectedFileName: 'single-child',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						attributes: [{ name: 'aA', value: 'value aA' }],
-					},
-				],
+			'child with attribute': {
+				sourceXml: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"><A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A></Root>`,
+				expectedQueries: ['//default:Root/default:A[@aA="value aA"]'],
 			},
-			{
-				description: 'namespace element with text',
-				fileContent: `
-					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${XMLNS_EXT_NAMESPACE}  ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
+			'namespace element with text': {
+				sourceXml: `
+					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${XMLNS_EXT_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
 						<A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A>
 						<ext:AA_3 aAA_3="value aAA_3" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3">text value</ext:AA_3>
 					</Root>
 				`,
-				expectedFileName: 'namespace-with-text',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						attributes: [{ name: 'aA', value: 'value aA' }],
-					},
-					{
-						id: '3',
-						tagName: 'AA_3',
-						namespace: DIALECTE_NAMESPACES.ext,
-						attributes: [{ name: 'aAA_3', value: 'value aAA_3' }],
-						value: 'text value',
-					},
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:Root/ext:AA_3[@aAA_3="value aAA_3"]',
+					'//default:Root/ext:AA_3[text()="text value"]',
 				],
 			},
-			{
-				description: 'multiple siblings',
-				fileContent: `
+			'multiple siblings': {
+				sourceXml: `
 					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
 						<A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A>
 						<B aB="value aB" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3"></B>
 					</Root>
 				`,
-				expectedFileName: 'multiple-siblings',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						attributes: [{ name: 'aA', value: 'value aA' }],
-					},
-					{
-						id: '3',
-						tagName: 'B',
-						attributes: [{ name: 'aB', value: 'value aB' }],
-					},
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:Root/default:B[@aB="value aB"]',
 				],
 			},
-			{
-				description: 'parent/child relationships',
-				fileContent: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"><A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"><AA_1 aAA_1="value aAA_1" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3"/></A></Root>`,
-				expectedFileName: 'parent-relationships',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-						children: [{ id: '2', tagName: 'A' }],
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						attributes: [{ name: 'aA', value: 'value aA' }],
-						parent: { id: '1', tagName: 'Root' },
-						children: [{ id: '3', tagName: 'AA_1' }],
-					},
-					{
-						id: '3',
-						tagName: 'AA_1',
-						attributes: [{ name: 'aAA_1', value: 'value aAA_1' }],
-						parent: { id: '2', tagName: 'A' },
-					},
+			'parent/child relationships': {
+				sourceXml: `<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1"><A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"><AA_1 aAA_1="value aAA_1" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3"/></A></Root>`,
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:Root/default:A/default:AA_1[@aAA_1="value aAA_1"]',
 				],
 			},
-			// TODO: Uncomment when issue #789 is resolved (filtering xmlns at import)
-			// {
-			// 	description: 'xmlns NOT imported as attributes',
-			// 	fileName: 'xmlns-not-imported.xml',
-			// 	fileContent: `
-			// 		<Root xmlns="http://root.org/ROOT" xmlns:ext="http://example.com/ext" version="1.0" ${DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
-			// 			<A name="TestA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A>
-			// 		</Root>
-			// 	`,
-			// 	expectedFileName: 'xmlns-not-imported',
-			// 	expectedRecords: [
-			// 		{
-			// 			id: '1',
-			// 			tagName: 'Root',
-			// 			// xmlns and xmlns:ext should NOT be in attributes
-			// 			// only version should be imported
-			// 			attributes: [{ name: 'version', value: '1.0' }],
-			// 			children: [{ id: '2', tagName: 'A' }],
-			// 		},
-			// 		{
-			// 			id: '2',
-			// 			tagName: 'A',
-			// 			attributes: [{ name: 'name', value: 'TestA' }],
-			// 			parent: { id: '1', tagName: 'Root' },
-			// 		},
-			// 	],
-			// },
-			{
-				description: 'qualified attributes with local name',
-				fileContent: `
+			'qualified attributes with local name': {
+				sourceXml: `
 					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${XMLNS_EXT_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
 						<A aA="value aA" ext:cA="value cA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2"></A>
 					</Root>
 				`,
-				expectedFileName: 'qualified-attributes',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-						children: [{ id: '2', tagName: 'A' }],
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						parent: { id: '1', tagName: 'Root' },
-						attributes: [
-							{ name: 'aA', value: 'value aA' },
-							{
-								name: 'cA',
-								value: 'value cA',
-								namespace: DIALECTE_NAMESPACES.ext,
-							},
-						],
-					},
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:Root/default:A[@ext:cA="value cA"]',
 				],
 			},
-			{
-				description: 'deep nesting',
-				fileContent: `
+			'deep nesting': {
+				sourceXml: `
 					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
 						<A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2">
 							<AA_1 aAA_1="value aAA_1" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3">
@@ -209,67 +80,56 @@ describe('Import', () => {
 						</A>
 					</Root>
 				`,
-				expectedFileName: 'deep-nesting',
-				expectedRecords: [
-					{
-						id: '1',
-						tagName: 'Root',
-					},
-					{
-						id: '2',
-						tagName: 'A',
-						attributes: [{ name: 'aA', value: 'value aA' }],
-					},
-					{
-						id: '3',
-						tagName: 'AA_1',
-						attributes: [{ name: 'aAA_1', value: 'value aAA_1' }],
-					},
-					{
-						id: '4',
-						tagName: 'AAA_1',
-						attributes: [{ name: 'aAAA_1', value: 'value aAAA_1' }],
-					},
-					{
-						id: '5',
-						tagName: 'AAAA_1',
-						attributes: [{ name: 'aAAAA_1', value: 'value aAAAA_1' }],
-					},
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:A/default:AA_1[@aAA_1="value aAA_1"]',
+					'//default:AA_1/default:AAA_1[@aAAA_1="value aAAA_1"]',
+					'//default:AAA_1/default:AAAA_1[@aAAAA_1="value aAAAA_1"]',
 				],
 			},
-		]
+		}
 
-		testCases.forEach((testCase) => {
-			testFeature(testCase)
-			testFeature(testCase) // test overwrite feature
+		runTestCases.withExport({
+			testCases,
+			act: async ({ source }): Promise<ActResult> => {
+				return { assertDatabaseName: source.databaseName }
+			},
 		})
+	})
 
-		function testFeature({ description, fileContent, expectedRecords }: TestCase) {
-			it(description, async () => {
-				const filename = `${description.replace(/\s+/g, '-')}-${crypto.randomUUID()}.xml`
-				const file = new File([fileContent], filename, {
-					type: 'text/plain',
+	describe('Overwrite', () => {
+		type TestCase = BaseXmlTestCase
+
+		const testCases: Record<string, TestCase> = {
+			'reimporting same XML overwrites without duplication': {
+				sourceXml: `
+					<Root ${XMLNS_DEFAULT_NAMESPACE} ${XMLNS_DEV_NAMESPACE} ${CUSTOM_RECORD_ID_ATTRIBUTE}="1">
+						<A aA="value aA" ${CUSTOM_RECORD_ID_ATTRIBUTE}="2">
+							<AA_1 aAA_1="value aAA_1" ${CUSTOM_RECORD_ID_ATTRIBUTE}="3"/>
+						</A>
+					</Root>
+				`,
+				expectedQueries: [
+					'//default:Root/default:A[@aA="value aA"]',
+					'//default:A/default:AA_1[@aAA_1="value aAA_1"]',
+				],
+				unexpectedQueries: ['//default:Root/default:A[2]'],
+			},
+		}
+
+		runTestCases.withExport({
+			testCases,
+			act: async ({ source, testCase }): Promise<ActResult> => {
+				const file = new File([testCase.sourceXml], `${source.databaseName}.xml`, {
+					type: 'text/xml',
 				})
-
-				const databaseNames = await importXmlFiles({
+				await importXmlFiles({
 					files: [file],
 					dialecteConfig: TEST_DIALECTE_CONFIG,
 					useCustomRecordsIds: true,
 				})
-				const databaseName = filename.replace(/\.xml$/, '')
-				expect(databaseNames).toContain(databaseName)
-
-				const databaseInstance = new Dexie(databaseName) as AnyDatabaseInstance
-				await databaseInstance.open()
-				await handleExpectedRecords({
-					dialecteConfig: TEST_DIALECTE_CONFIG,
-					databaseInstance,
-					expectedRecords,
-				})
-
-				databaseInstance.close()
-				databaseNames.push(databaseName)
-			})
-		}
+				return { assertDatabaseName: source.databaseName }
+			},
+		})
 	})
 })